pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gql_defrag"
version = "0.0.1"
authors = [
  { name="Jelle Zijlstra", email="jelle@quora.com" },
]
description = "Merge GraphQL fragments into the parent query"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "graphql-core==2.3.2"
]

[project.urls]
"Homepage" = "https://github.com/JelleZijlstra/gql-defrag"

[tool.pyright]
pythonVersion = "3.9"

[tool.black]
target-version = ['py39']
include = '\.pyi?$'
skip-magic-trailing-comma = true

[tool.pycln]
all = true

[tool.isort]
profile = "black"
combine_as_imports = true
skip_gitignore = true
